Divide by zero fix, fat entry root directory listing.
Additional checks for wrong bpb settings in order to avoid divide by
zero problems. The root directory listing will be printed and some logic
is involved regarding appending the extension part separated by `.'.
Files that do not have an extension are not compatible with FAT so we
don't care about those.