FAT root directory entries pretty printed in table.
FAT root directory entries are printed in a table containing all
information available in the entry. Long filenames are not supported and
we are not sure whether we are going to support it for floppy disks,
since most practical use cases for FAT12 do not require LFN.

Some filenames do not have an extension because they are directories.
Menuet OS also seems to use filenames without extensions for executable
files. I don't know if this is Menuet specific and if this violates the
FAT specification, but I don't care about that special case.