FAT12 initial program loader
Initial program loader that fetches next stage from FAT12 filesystem and
jumps to it. Most formulas are computed on the fly and can be changed in
macro definitions. Lots of dirty comments have to be removed once the
code has been refactored/cleaned up.

Tested with the MikeOS kernel and worked perfectly.