poke to address
Press `a' to start poking to address. You need to specify both the
segment value and the relative address in that particular segment. E.g.:

a 0800: 0000, ca, fe, ba, be,

Press space to stop poking. Currently, only one byte at a time can be
poked into memory and backspace is not supported.