dump memory
Press `d' to start dumping from address. You need to specify both the
segment value and the relative address in that particular segment. E.g.:

d 0800: 0000* 04, CA, FE, BA, BE

The maximum number of bytes that can be printed is 255. If you enter 0,
it will print 65536 characters since it underflows.